pyproject.toml
[tool.poetry]
name = "monitoring-bot"
version = "0.1.0"
description = ""
authors = ["eakotelnikov <eakotelnikov@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "13.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
